KNOX-1996: More code review fixes. Creating base class for WebsocketEcho and WebsocketEcho1 tests
